ai/api/fast_api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.2
httpx==0.28.1
python-dotenv==1.1.0
Pillow==11.1.0
pdfplumber==0.11.9